Moved utils and consts from pkg __init__ to new files.
