+++ core/consts.py
@@ -1,0 +1,7 @@
+FILE_DOC_FORMAT = """~
+Name: {0}
+Purpose: {1}
+Author: DavidGuler (davidguler.1x@gmail.com)
+~"""
+PACKAGE_PATH = os.path.dirname(__file__)
+REGEX_FOR_DOC = re.compile("~\n")
